refactor project. convert to mvn project.
