+++ Demonstration/src/jLQNInterface/src/main/java/ca/appsimulations/jlqninterface/algorithms/Algorithm.java
