+++ Demonstration/src/jLQNInterface/src/main/java/ca/appsimulations/jlqninterface/core/Configuration.java
