+++ Demonstration/src/jLQNInterface/src/main/java/ca/appsimulations/jlqninterface/drivers/Driver.java
@@ -10,16 +10,6 @@ package ca.appsimulations.jlqninterface.drivers;
  * This is the main Driver program. Please run this program first.
  */
 
-import java.io.IOException;
-import java.util.ArrayList;
-
-import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.parsers.SAXParser;
-import javax.xml.parsers.SAXParserFactory;
-
-import org.xml.sax.SAXException;
-import org.xml.sax.XMLReader;
-
 import ca.appsimulations.jlqninterface.algorithms.Algorithm;
 import ca.appsimulations.jlqninterface.algorithms.Algorithm1;
 import ca.appsimulations.jlqninterface.core.Model;
