+++ Demonstration/src/jLQNInterface/src/main/java/ca/appsimulations/jlqninterface/lqns/entities/EntryPhaseActivities.java
