+++ Demonstration/src/jLQNInterface/src/main/java/ca/appsimulations/jlqninterface/lqns/modelhandling/LQNXMLAttributes.java
