+++ Demonstration/src/jLQNInterface/src/main/java/ca/appsimulations/jlqninterface/lqns/modelhandling/LQNXmlResultParser.java
@@ -1,17 +1,13 @@
 package ca.appsimulations.jlqninterface.lqns.modelhandling;
 
-import org.xml.sax.Attributes;
-import org.xml.sax.SAXException;
-
 import ca.appsimulations.jlqninterface.core.Model;
 import ca.appsimulations.jlqninterface.lqns.entities.ActivityDef;
 import ca.appsimulations.jlqninterface.lqns.entities.ActivityPhases;
+import org.xml.sax.Attributes;
+import org.xml.sax.SAXException;
+
 import ca.appsimulations.jlqninterface.lqns.entities.EntryAcType;
-import ca.appsimulations.jlqninterface.lqns.entities.LQNConstants;
-import ca.appsimulations.jlqninterface.lqns.entities.LQNDefaults;
-import ca.appsimulations.jlqninterface.lqns.entities.ProcessorSchedulingType;
 import ca.appsimulations.jlqninterface.lqns.entities.SynchCall;
-import ca.appsimulations.jlqninterface.lqns.entities.TaskSchedulingType;
 
 /**
  * @author Yasir Shoaib (2011,2012)
@@ -21,11 +17,12 @@ import ca.appsimulations.jlqninterface.lqns.entities.TaskSchedulingType;
  * Some LQN classes and their members are outlined as UML class diagrams in LQNS User Manual.
  * For details regarding these LQN classes and members refer to LQNS User Manual.
  */
-public class LQNXmlModelInputParser extends LQNParser {
+public class LQNXmlResultParser extends LQNParser {
 
-	public LQNXmlModelInputParser(Model workspace) {
+	public LQNXmlResultParser(Model workspace) {
 		super(workspace);
 	}
+	
 	
 	@Override
 	public void startElement(String uri, String localName, String qName, Attributes attributes) throws SAXException {
@@ -42,10 +39,18 @@ public class LQNXmlModelInputParser extends LQNParser {
 		String attrReplication = attributes.getValue(LQNXMLAttributes.REPLICATION.toString());
 
 		switch (et) {
+			case RESULT_GENERAL:
+
+				String attrValid = attributes.getValue(LQNXMLAttributes.VALID.toString());
+				workspace.getResult().setResultValid(attrValid);
+				break;
+				
 			case PROCESSOR:
+				
 				// name
 				curProcessor = workspace.getProcessorByName(attrName, true);
-
+				
+				/*
 				// scheduling
 				ProcessorSchedulingType psType = ProcessorSchedulingType.getValue(attrScheduling);
 				curProcessor.setScheduling(psType);
@@ -65,20 +70,21 @@ public class LQNXmlModelInputParser extends LQNParser {
 				} else if ((attrReplication == null)) {
 					curProcessor.setReplication(LQNDefaults.PROCESSOR_REPLICATION.getValue());
 				}
+				*/
 				break;
 			case RESULT_PROCESSOR:
-/*
+
 				// utilization
 				String attrUtilization = attributes.getValue(LQNXMLAttributes.UTILIZATION.toString());
 				float utilization = Float.parseFloat(attrUtilization);
 				curProcessor.getResult().setUtilization(utilization);
-				*/
 				break;
-				
 			case TASK:
+				
 				// name
 				curTask = workspace.getTaskByName(attrName, curProcessor, true);
-
+				
+				/*
 				// scheduling
 				TaskSchedulingType schedulingType = TaskSchedulingType.getValue(attrScheduling);
 				curTask.setScheduling(schedulingType);
@@ -102,10 +108,10 @@ public class LQNXmlModelInputParser extends LQNParser {
 				} else if ((attrReplication == null)) {
 					curTask.setReplication(LQNDefaults.TASK_REPLICATION.getValue());
 				}
-
+				*/
+				
 				break;
 			case RESULT_TASK:
-				/*
 				attrUtilization = attributes.getValue(LQNXMLAttributes.UTILIZATION.toString());
 				String attrThroughput = attributes.getValue(LQNXMLAttributes.THROUGHPUT.toString());
 				String attrPh1Utilization = attributes.getValue(LQNXMLAttributes.PHASE1_UTILIZATION.toString());
@@ -126,18 +132,15 @@ public class LQNXmlModelInputParser extends LQNParser {
 				curTask.getResult().setUtilization(utilization);
 				curTask.getResult().setPhase1_utilization(ph1Utilization);
 				curTask.getResult().setProc_utilization(procUtilization);
-				*/
 				break;
-				
 			case ENTRY:
 				String attrType = attributes.getValue(LQNXMLAttributes.TYPE.toString());
 
 				// name,type
 				curEntry = workspace.getEntryByName(attrName, curTask, true);
-				curEntry.setEntryType(EntryAcType.getValue(attrType));
+				//curEntry.setEntryType(EntryAcType.getValue(attrType));
 				break;
 			case RESULT_ENTRY:
-				/*
 				attrUtilization = attributes.getValue(LQNXMLAttributes.UTILIZATION.toString());
 				attrThroughput = attributes.getValue(LQNXMLAttributes.THROUGHPUT.toString());
 				String attrSquaredCoeff = attributes.getValue(LQNXMLAttributes.SQUARED_COEFF_VARIATION.toString());
@@ -171,17 +174,15 @@ public class LQNXmlModelInputParser extends LQNParser {
 				curEntry.getResult().setUtilization(utilization);
 				curEntry.getResult().setSquared_coeff_variation(squaredCoeffVariation);
 				curEntry.getResult().setProc_utilization(procUtilization);
-				*/
-				
+
 				break;
-			
 			case ENTRY_PHASE_ACTIVITIES:
-				curEntry.generateEntryPhaseActivities();
+				//curEntry.generateEntryPhaseActivities();
 				isEntryPhaseActivities = true;
 				break;
 			case TASK_ACTIVITIES:
 				// TODO need to find task activity by name
-				curTaskActivities = curTask.generateTaskActivities();
+				//curTaskActivities = curTask.generateTaskActivities();
 				isTaskActivities = true;
 				break;
 			case ACTIVITY:
@@ -216,14 +217,15 @@ public class LQNXmlModelInputParser extends LQNParser {
 				// "\t\t taskName: " + curTask + "\t\t entryName: " + curEntry);
 
 				// host-demand-mean
-				if (attrHostDemand != null) {
+				
+				/*if (attrHostDemand != null) {
 					float hostDemand = Float.parseFloat(attrHostDemand);
 					curActivity.setHost_demand_mean(hostDemand);
 				}
-
+				*/
+				
 				break;
 			case RESULT_ACTIVITY:
-				/*
 				String attrProcWaiting = attributes.getValue(LQNXMLAttributes.PROC_WAITING.toString());
 				String attrServiceTime = attributes.getValue(LQNXMLAttributes.SERVICE_TIME.toString());
 				String attrServiceTimeVar = attributes.getValue(LQNXMLAttributes.SERVICE_TIME_VARIANCE.toString());
@@ -239,9 +241,8 @@ public class LQNXmlModelInputParser extends LQNParser {
 				curActivity.getResult().setService_time(serviceTime);
 				curActivity.getResult().setService_time_variance(serviceTimeVar);
 				curActivity.getResult().setUtilization(utilization);
-				*/
+
 				break;
-				
 			case SYNCH_CALL:
 				// dest, callsmean, fanin,fanout
 				String attrDest = attributes.getValue(LQNXMLAttributes.DEST.toString());
@@ -269,7 +270,8 @@ public class LQNXmlModelInputParser extends LQNParser {
 						ad.addSynchCall(s);
 					}
 				}
-
+				
+				/*
 				if (attrFanin != null) {
 					s.setFanin(Integer.parseInt(attrFanin));
 				}
@@ -277,6 +279,7 @@ public class LQNXmlModelInputParser extends LQNParser {
 				if (attrFanout != null) {
 					s.setFanout(Integer.parseInt(attrFanout));
 				}
+				*/
 
 				/*
 				 * String attrDest =
@@ -359,4 +362,5 @@ public class LQNXmlModelInputParser extends LQNParser {
 				break;
 		}
 	}
+
 }
